Fixed bug regarding premature exit from editing series
If user exited editing series without saving and then edited the series again, program threw a null reference. This was because the ReceiveSeriesData method (which fires when the Edit view received a message) would delete the entry in the database. Entry is now deleted on when user saves their changes.