Fixed #8
Problem actually arose in an assumption made regarding track data. On most tracks, track_name was the same as the name displayed in the result file. However, there was some situations in which the registered trademark symbol did not get placed in the result file. Probably something to do with encoding. Program now ignores trademark symbols entirely.